test: fix plan teardown in roby app tests (non-Self)
